More float conversions
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-experimental@2750 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/builders/residfp-builder/residfp/resample/SincResampler.h
@@ -68,16 +68,16 @@ private:
     /// Filter length
     int firN;
 
-    const int cyclesPerSample;
+    const double cyclesPerSample;
 
-    int sampleOffset;
+    double sampleOffset;
 
     float outputValue;
 
     float sample[RINGSIZE * 2];
 
 private:
-    float fir(int subcycle);
+    float fir(float subcycle);
 
 public:
     /**
